Slightly change procedure.

+++ src/global.h
@@ -358,7 +358,7 @@ struct edge {
 	int xtid;	// Index of edge in binary indexed tree
 	int idx;	// Index of the edge in eptr array of population
 	int trunk_id;	// Index in trunk genealogy (used by merge_floating)
-	double ub;
+	int ub;
 };
 
 #endif
